Refactor code into files
