+++ src/modules/users/controllers/index.ts
@@ -1,0 +1,1 @@
+export * from './users.controller';
