+++ src/modules/users/services/index.ts
@@ -1,0 +1,1 @@
+export * from './users.service';
